add fixes
