+++ vision_agent_tools/shared_types.py
@@ -88,3 +88,39 @@ class BboxAndMaskLabel(BboxLabel):
     class Config:
         arbitrary_types_allowed = True
         populate_by_name = True
+
+
+class FlorenceV2ODRes(BaseModel):
+    """
+    Schema for the <OD> task.
+    """
+
+    bboxes: list[BoundingBox] = Field(
+        ..., description="list of bounding boxes, each represented as [x1, y1, x2, y2]"
+    )
+    labels: list[str] = Field(
+        ..., description="list of labels corresponding to each bounding box"
+    )
+
+    class Config:
+        schema_extra = {
+            "example": {
+                "<OD>": {
+                    "bboxes": [
+                        [
+                            33.599998474121094,
+                            159.59999084472656,
+                            596.7999877929688,
+                            371.7599792480469,
+                        ],
+                        [
+                            454.0799865722656,
+                            96.23999786376953,
+                            580.7999877929688,
+                            261.8399963378906,
+                        ],
+                    ],
+                    "labels": ["car", "door"],
+                }
+            }
+        }
